Refactoring of Payment Module and Modal
